+++ packages/ilmomasiina-frontend/src/components/Header/index.tsx
@@ -4,9 +4,9 @@ import { Button, Container, Navbar } from "react-bootstrap";
 import { useTranslation } from "react-i18next";
 import { Link } from "react-router-dom";
 
-import logo from "../../assets/logo.svg";
-import branding from "../../branding";
+import defaultLogo from "../../assets/logo.svg";
 import i18n from "../../i18n";
+import { useHeaderBranding } from "../../modules/branding";
 import paths from "../../paths";
 
 import "./Header.scss";
@@ -19,14 +19,15 @@ const Header = () => {
     i18n: { language },
     t,
   } = useTranslation();
+  const { headerTitle, headerTitleShort, logo } = useHeaderBranding();
 
   return (
     <Navbar>
       <Container className="gap-sm-2">
         <Link to={paths.eventsList} className="navbar-brand">
-          <img className="navbar-logo" src={logo} alt="Logo" />
-          <span className="d-none d-sm-inline">{branding.headerTitle}</span>
-          <span className="d-sm-none">{branding.headerTitleShort}</span>
+          <img className={logo ? "navbar-logo navbar-logo-custom" : "navbar-logo"} src={logo ?? defaultLogo} alt="Logo" />
+          <span className="d-none d-sm-inline">{headerTitle}</span>
+          <span className="d-sm-none">{headerTitleShort}</span>
         </Link>
         {language !== "fi" && (
           <Button onClick={() => i18n.changeLanguage("fi")}>{t("header.switchLanguage", { lng: "fi" })}</Button>